@@ -1,5 +1,5 @@
 {
-    "name": "frontend",
+    "name": "simpleshub-frontend",
     "version": "1.0.0",
     "private": true,
     "scripts": {
@@ -8,20 +8,22 @@
         "lint": "vue-cli-service lint"
     },
     "dependencies": {
-        "axios": "^0.21.1",
-        "core-js": "^3.6.5",
-        "vue": "^2.6.14"
+        "axios": "^1.4.0",
+        "bootstrap": "^5.3.3",
+        "bootstrap-vue": "^2.23.1",
+        "bootstrap-vue-3": "^0.3.2",
+        "core-js": "^3.8.3",
+        "vue": "^3.3.4"
     },
     "devDependencies": {
-        "@vue/cli-plugin-babel": "^4.5.13",
-        "@vue/cli-plugin-eslint": "^4.5.13",
-        "@vue/cli-service": "^4.5.13",
-        "babel-eslint": "^10.1.0",
-        "eslint": "^7.32.0",
-        "eslint-plugin-vue": "^7.20.0", 
-        "sass": "^1.32.0",
-        "sass-loader": "^10.1.1",
-        "vue-template-compiler": "^2.6.14"
+        "@vue/cli-plugin-babel": "^5.0.8",
+        "@vue/cli-plugin-eslint": "^5.0.8",
+        "@vue/cli-service": "^5.0.8",
+        "@vue/compiler-sfc": "^3.3.4",
+        "eslint": "^8.49.0",
+        "eslint-plugin-vue": "^9.17.0",
+        "sass": "^1.63.6",
+        "sass-loader": "^13.3.2"
     },
     "browserslist": [
         "> 1%",
