@@ -1,15 +1,10 @@
-import Vue from 'vue';
+import { createApp } from 'vue';
 import App from './App.vue';
-import { BootstrapVue, IconsPlugin } from 'bootstrap-vue'; // Importa o BootstrapVue
+import { BootstrapVue3 } from 'bootstrap-vue-3';
 
-import 'bootstrap/dist/css/bootstrap.css';
-import 'bootstrap-vue/dist/bootstrap-vue.css';
+import 'bootstrap/dist/css/bootstrap.min.css';
+import 'bootstrap-vue-3/dist/bootstrap-vue-3.css';
 
-Vue.use(BootstrapVue);
-Vue.use(IconsPlugin);
-
-Vue.config.productionTip = false;
-
-new Vue({
-  render: h => h(App),
-}).$mount('#app');
+const app = createApp(App);
+app.use(BootstrapVue3);
+app.mount('#app');
