@@ -7,7 +7,7 @@ class UploadController {
       const filePath = req.file.path;
       const cpfs = await PdfService.extractCpfFromPdf(filePath);
       await CpfService.saveCpfs(cpfs);
-      res.status(200).json({ cpfs });
+      res.status(200).json({ msg: "Cpfs enviados com sucesso", statusCode: 200 });
     } catch (error) {
       res.status(500).json({ error: 'Erro ao processar o PDF' });
     }
