@@ -1,17 +1,18 @@
 module.exports = {
-    root: true,
-    env: {
-      node: true,
-    },
-    extends: [
-      'plugin:vue/essential', 
-      'eslint:recommended',    
-    ],
-    parserOptions: {
-      parser: 'babel-eslint',  
-    },
-    rules: {
-      'no-console': 'off',     
-      'no-debugger': 'off',    
-    },
-  };
+  root: true,
+  env: {
+    node: true,
+  },
+  extends: [
+    'plugin:vue/essential', 
+    'eslint:recommended',    
+  ],
+  parserOptions: {
+    parser: '@babel/eslint-parser',  
+    requireConfigFile: false,  
+  },
+  rules: {
+    'no-console': 'off',     
+    'no-debugger': 'off',    
+  },
+};
